# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "FASE 2 API",
        "description": "FASE 2 Swagger OpenApi",
        "contact": {
            "email": "eaguilar.arrezola@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0",
        "x-logo": {
            "url": "https://via.placeholder.com/190x90.png?text=L5-Swagger"
        }
    },
    "paths": {
...
        "/api/providers": {
            "get": {
                "tags": [
                    "provides"
                ],
                "summary": "Get all providers",
                "operationId": "ffc5716dada584c9dc1415e8e7108f1a",
                "responses": {
                    "200": {
                        "description": "Valida existencia de usuario."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "provides"
                ],
                "summary": "Add provider",
                "operationId": "a28875d77542c226463fa69c09503859",
                "responses": {
                    "200": {
                        "description": "Valida existencia de usuario."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/providers/{id}": {
            "get": {
                "tags": [
                    "provides"
                ],
                "summary": "Get provider",
                "operationId": "f246e498b36131a55996564f68b97495",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Valida existencia de usuario."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "provides"
                ],
                "summary": "Update provider",
                "operationId": "dd704e72367f2f4356d0ea0d1f210190",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Valida existencia de usuario."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "provides"
                ],
                "summary": "Delete provider",
                "operationId": "504d3543df3e692dec22c6d86e6c1984",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Valida existencia de usuario."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearer_token": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}